Episode Model and loading fixes
fixed some items not loading

+++ android-client/app/src/main/java/app/kyushu/aynime/data/remote/gogo_flask/responses/Episode.kt
@@ -2,6 +2,6 @@ package app.kyushu.aynime.data.remote.gogo_flask.responses
 
 data class Episode(
     val id: String,
-    val number: Int,
+    val number: Float,
     val sub: Boolean
 )
